+++ ChangeLog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Version [0.7.0](https://github.com/MC2IT/Agicap.net/compare/v0.6.0...v0.7.0)
+- **Suppliers:** added the `LegalAddress` class.
+
 ## Version [0.6.0](https://github.com/MC2IT/Agicap.net/compare/v0.5.0...v0.6.0)
 - Removed the `NestedList` class.
 - **Chart of accounts:** added the accounting account and third-party APIs.